Extract interface from Form class
